fix(ab): circular dependency
